chore: make compitable the account generator with custome
